+++ src/admin/client/modules/productCategories/components/list.tsx
@@ -3,52 +3,7 @@ import FontIcon from "material-ui/FontIcon"
 import { List, ListItem } from "material-ui/List"
 import React from "react"
 import { Link } from "react-router-dom"
-
-const styles = {
-  selectedItem: {
-    backgroundColor: "rgba(0, 0, 0, 0.1)",
-  },
-  innerItem: {
-    paddingLeft: 55,
-  },
-  nestedListStyle: {
-    padding: "0 0 0 15px",
-  },
-}
-
-const FolderIcon = <FontIcon className="material-icons">folder</FontIcon>
-const DraftIcon = <FontIcon className="material-icons">visibility_off</FontIcon>
-
-class Item extends React.PureComponent {
-  constructor(props) {
-    super(props)
-  }
-
-  handleClick = () => {
-    const { item } = this.props
-    this.props.onSelect(item.id)
-  }
-
-  render() {
-    const { item, opened, selectedId, nestedItems } = this.props
-    const icon = item.enabled ? FolderIcon : DraftIcon
-    const style = item.id === selectedId ? styles.selectedItem : null
-
-    return (
-      <ListItem
-        className="treeItem"
-        initiallyOpen={opened}
-        style={style}
-        innerDivStyle={styles.innerItem}
-        primaryText={item.name}
-        nestedItems={nestedItems}
-        leftIcon={icon}
-        onClick={this.handleClick}
-        nestedListStyle={styles.nestedListStyle}
-      />
-    )
-  }
-}
+import Item, { styles } from "./item"
 
 class Categories extends React.Component {
   constructor(props) {
